@@ -51,52 +47,47 @@
         ]
     },
     "devDependencies": {
-        "@babel/core": "^7.20.12",
-        "@babel/plugin-transform-runtime": "^7.19.6",
-        "@babel/preset-env": "^7.20.2",
+        "@babel/core": "^7.21.4",
+        "@babel/plugin-transform-runtime": "^7.21.4",
+        "@babel/preset-env": "^7.21.4",
         "@babel/preset-react": "^7.18.6",
-        "@babel/preset-typescript": "^7.18.6",
-        "@babel/runtime": "^7.20.7",
+        "@babel/preset-typescript": "^7.21.4",
+        "@babel/runtime": "^7.21.0",
         "@types/fork-ts-checker-webpack-plugin": "^0.4.5",
-        "@types/node": "^18.11.18",
-        "@types/react": "^18.0.26",
-        "@types/react-dom": "^18.0.10",
+        "@types/node": "^18.15.11",
+        "@types/react": "^18.0.35",
+        "@types/react-dom": "^18.0.11",
         "@types/styled-components": "^5.1.26",
-        "@types/webpack": "^5.28.0",
+        "@types/webpack": "^5.28.1",
         "@types/webpack-dev-server": "^4.7.2",
-        "@typescript-eslint/eslint-plugin": "^5.48.1",
-        "@typescript-eslint/parser": "^5.48.1",
+        "@typescript-eslint/eslint-plugin": "^5.58.0",
+        "@typescript-eslint/parser": "^5.58.0",
         "babel-loader": "^9.1.2",
-        "babel-plugin-styled-components": "^2.0.7",
-        "clean-webpack-plugin": "^4.0.0",
-        "copy-webpack-plugin": "^9.0.1",
-        "cross-env": "^7.0.3",
-        "css-loader": "^5.2.7",
-        "css-minimizer-webpack-plugin": "^4.2.2",
-        "eslint": "^8.32.0",
-        "eslint-config-airbnb": "^19.0.4",
-        "eslint-config-prettier": "^8.6.0",
+        "babel-plugin-styled-components": "^2.1.1",
+        "case-sensitive-paths-webpack-plugin": "^2.4.0",
+        "copy-webpack-plugin": "^11.0.0",
+        "css-loader": "^6.7.3",
+        "css-minimizer-webpack-plugin": "^5.0.0",
+        "eslint": "^8.38.0",
+        "eslint-config-prettier": "^8.8.0",
         "eslint-loader": "^4.0.2",
         "eslint-plugin-prettier": "^4.2.1",
-        "eslint-plugin-react": "^7.32.0",
+        "eslint-plugin-react": "^7.32.2",
         "eslint-plugin-react-hooks": "^4.6.0",
-        "eslint-webpack-plugin": "^3.2.0",
-        "fork-ts-checker-webpack-plugin": "^7.3.0",
+        "fork-ts-checker-webpack-plugin": "^8.0.0",
         "html-webpack-plugin": "^5.5.0",
         "husky": "^8.0.3",
         "install": "^0.13.0",
-        "less-loader": "^11.1.0",
-        "lint-staged": "^13.1.0",
-        "mini-css-extract-plugin": "^2.7.2",
-        "prettier": "^2.8.3",
-        "style-loader": "^3.3.1",
+        "lint-staged": "^13.2.1",
+        "mini-css-extract-plugin": "^2.7.5",
+        "prettier": "^2.8.7",
+        "style-loader": "^3.3.2",
         "ts-loader": "^9.4.2",
-        "ts-node": "^10.9.1",
-        "tsconfig-paths": "^4.1.2",
-        "typescript": "^4.9.4",
-        "webpack": "^5.75.0",
+        "tsconfig-paths": "^4.2.0",
+        "typescript": "^5.0.4",
+        "webpack": "^5.78.0",
         "webpack-cli": "^5.0.1",
-        "webpack-dev-server": "^4.11.1"
+        "webpack-dev-server": "^4.13.2"
     },
     "resolutions": {
         "@types/react": "^18.x",
