@@ -1,62 +1,69 @@
-const path = require('path');
-const webpack = require('webpack');
-const HtmlWebpackPlugin = require('html-webpack-plugin');
-const TypeScriptTypeChecker = require('fork-ts-checker-webpack-plugin');
-const CssMinimizerPlugin =require('css-minimizer-webpack-plugin');
+import path from 'path';
+import webpack from 'webpack';
+import { mergeWithCustomize } from 'webpack-merge';
+import common from './webpack.common';
+import HtmlWebpackPlugin from 'html-webpack-plugin';
+import TypeScriptTypeChecker from 'fork-ts-checker-webpack-plugin';
+import CssMinimizerPlugin from 'css-minimizer-webpack-plugin';
 
-module.exports = {
-    entry: {
-        'familie-ef-ettersending': ['./src/frontend/index.tsx'],
-    },
-    devtool: 'source-map',
-    resolve: {
-        extensions: ['.ts', '.tsx', '.js', '.jsx', '.json', '.less'],
-    },
-    module: {
-        rules: [
-            {
-                test: /\.(ts|tsx)$/,
-                enforce: 'pre',
-                exclude: /node_modules/,
-                use: [
-                    {
-                        options: {
-                            eslintPath: require.resolve('eslint'),
-                        },
-                        loader: require.resolve('eslint-loader'),
-                    },
-                ],
+const config = mergeWithCustomize({
+  'entry.familie-ef-ettersending': 'prepend',
+  'module.rules': 'append',
+})(common, {
+  entry: {
+    'familie-ef-ettersending': ['./src/frontend/index.tsx'],
+  },
+  devtool: 'source-map',
+  resolve: {
+    extensions: ['.ts', '.tsx', '.js', '.jsx', '.json'],
+  },
+  module: {
+    rules: [
+      {
+        test: /\.(ts|tsx)$/,
+        enforce: 'pre',
+        exclude: /node_modules/,
+        use: [
+          {
+            options: {
+              eslintPath: 'eslint',
             },
-            {
-                test: /\.(js|ts|tsx)$/,
-                loader: 'ts-loader',
-                exclude: /node_modules/,
-                options: {
-                    transpileOnly: true,
-                },
-            },
+            loader: 'eslint-loader',
+          },
         ],
-    },
-    optimization: {
-        splitChunks: {
-            cacheGroups: {
-                vendor: {
-                    test: /[\\/]node_modules[\\/]/,
-                    name: 'vendors',
-                    chunks: 'all',
-                },
-            },
+      },
+      {
+        test: /\.(js|ts|tsx)$/,
+        loader: 'ts-loader',
+        exclude: /node_modules/,
+        options: {
+          transpileOnly: true,
         },
-        runtimeChunk: true,
-    },
-    plugins: [
-        new HtmlWebpackPlugin({
-            template: path.join(__dirname, '../../src/frontend/index.html'),
-            inject: 'body',
-            alwaysWriteToDisk: true,
-        }),
-        new TypeScriptTypeChecker(),
-        new webpack.NoEmitOnErrorsPlugin(),
-        new CssMinimizerPlugin(),
+      },
     ],
-};
+  },
+  optimization: {
+    splitChunks: {
+      cacheGroups: {
+        vendor: {
+          test: /[\\/]node_modules[\\/]/,
+          name: 'vendors',
+          chunks: 'all',
+        },
+      },
+    },
+    runtimeChunk: true,
+  },
+  plugins: [
+    new HtmlWebpackPlugin({
+      template: path.join(process.cwd(), '../../src/frontend/index.html'),
+      inject: 'body',
+      alwaysWriteToDisk: true,
+    }),
+    new TypeScriptTypeChecker(),
+    new webpack.NoEmitOnErrorsPlugin(),
+    new CssMinimizerPlugin(),
+  ],
+});
+
+export default config;
