@@ -1,12 +1,13 @@
 import CssMinimizerWebpackPlugin from 'css-minimizer-webpack-plugin';
-import MiniCssExtractPlugin from 'mini-css-extract-plugin';
 import TerserWebpackPlugin from 'terser-webpack-plugin';
-import webpack from 'webpack';
 import { CustomizeRule, mergeWithRules } from 'webpack-merge';
+import MiniCssExtractPlugin from 'mini-css-extract-plugin';
 
-import baseConfig, { createHtmlWebpackPlugin } from './webpack.common.config';
+import baseConfig, {
+  createHtmlWebpackPlugin,
+} from './webpack.common.config.js';
 
-const prodConfig: webpack.Configuration = mergeWithRules({
+const prodConfig = mergeWithRules({
   module: {
     rules: {
       test: CustomizeRule.Match,
@@ -35,11 +36,11 @@ const prodConfig: webpack.Configuration = mergeWithRules({
             loader: 'css-loader',
             options: {
               modules: {
-                compileType: 'icss',
+                mode: 'icss',
               },
+              importLoaders: 1,
             },
           },
-          { loader: 'less-loader' },
         ],
       },
     ],
